Moved spi package to dns.spi as per #742
