+++ gcloud-java-dns/src/main/java/com/google/gcloud/dns/Dns.java
@@ -20,7 +20,7 @@ import com.google.common.base.Joiner;
 import com.google.common.collect.Sets;
 import com.google.gcloud.Page;
 import com.google.gcloud.Service;
-import com.google.gcloud.spi.DnsRpc;
+import com.google.gcloud.dns.spi.DnsRpc;
 
 import java.io.Serializable;
 import java.util.Set;
